fix: refactored job rows to reduce component size
